Update dependencies, start adding room terrain support.
At this point, working on actual graphics code should be possible!

We still don't have a framework for caching data from websocket connections, but that can be done later.

Right now we should get displaying terrain grids for rooms working, and then getting map view and room detail on top of that sohuldn't be too hard.